feat: rootfs - build/upload action and Dockerfile
In order to facilitate Finch on Windows, we need a root filesystem. We
will use this Dockerfile as a basis for that root filesystem - using
`docker export` to turn a built container into an archived rootfs. For
the scope of these changes, create the Dockerfile and an action that
runs on changes to the file to build and push to ECR repo.

In order to create the container used as an intermediate step to export
its rootfs, we need to tell buildkit to load the image into Docker. This
exposes a limitation of buildkit to load multiplatform images, thus why
the action runs two build-rootfs-image jobs - one for each arch.

Additionally, network performance of the arm64 build is quite slow -
downloads of packages via dnf are on the scale of kb/s (but this has not
been observed on Ubuntu or Alpine images)

Signed-off-by: Gavin Inglis <giinglis@amazon.com>