ci: add script to update rootfs URL
for Finch on Windows runfinch/finch#492, we need to provide a rootfs to WSL2. This
rootfs lives in our dependencies bucket, and will be updated from
time-to-time for security patches, bug fixes, etc. This commit will
automatically pull the most recent rootfs from the depenedencies bucket
as part of the Update Deps action.

This commit also updates the upload rootfs action to use better pathing
than storing the rootfs at root level of S3 bucket.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>