ci: add script to update rootfs URL (#131)
Issue #, if available:

runfinch/finch#492

*Description of changes:*

for Finch on Windows runfinch/finch#492, we need to provide a rootfs to
WSL2. This rootfs lives in our dependencies bucket, and will be updated
from time-to-time for security patches, bug fixes, etc. This commit will
automatically pull the most recent rootfs from the depenedencies bucket
as part of the Update Deps action.

*Testing done:*

locally run script with FINCH_ROOTFS_URL ?=
https://deps.runfinch.com/finch-rootfs-production-amd64-000000000.tar.zst
set in Makefile, observe update.


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>